+++ nylas/client/client.py
@@ -50,8 +50,7 @@ class APIClient(json.JSONEncoder):
     def __init__(self, app_id=environ.get('NYLAS_APP_ID'),
                  app_secret=environ.get('NYLAS_APP_SECRET'),
                  access_token=environ.get('NYLAS_ACCESS_TOKEN'),
-                 api_server=API_SERVER,
-                 scopes=None):
+                 api_server=API_SERVER):
         if not api_server.startswith("https://"):
             raise Exception("When overriding the Nylas API server address, you"
                             " must include https://")
@@ -63,8 +62,6 @@ class APIClient(json.JSONEncoder):
 
         self.app_secret = app_secret
         self.app_id = app_id
-
-        self.scopes = scopes or ['email', 'calendar', 'contacts']
 
         self.session = requests.Session()
         self.version = __VERSION__
@@ -112,11 +109,16 @@ class APIClient(json.JSONEncoder):
             if 'Authorization' in self.session.headers:
                 del self.session.headers['Authorization']
 
-    def authentication_url(self, redirect_uri, login_hint='', state=''):
+    def authentication_url(self, redirect_uri, login_hint='', state='', scopes=None):
+        if not scopes:
+            scopes = ['email', 'calendar', 'contacts']
+        if isinstance(scopes, str):
+            scopes = [scopes]
+
         args = {'redirect_uri': redirect_uri,
                 'client_id': self.app_id or 'None',  # 'None' for back-compat
                 'response_type': 'code',
-                'scopes': ','.join(self.scopes),
+                'scopes': ','.join(scopes),
                 'login_hint': login_hint,
                 'state': state}