+++ tests/test_client.py
@@ -90,8 +90,7 @@ def test_client_authentication_url(api_client, api_url):
     assert urls_equal(expected3, actual3)
 
 
-def test_client_authentication_url_custom_scopes(api_client_email_only, api_url):
-    api_client = api_client_email_only
+def test_client_authentication_url_custom_scopes(api_client, api_url):
     expected = (
         URLObject(api_url)
         .with_path("/oauth/authorize")
@@ -104,20 +103,14 @@ def test_client_authentication_url_custom_scopes(api_client_email_only, api_url)
             ('scopes', 'email'),
         ])
     )
-    actual = URLObject(api_client.authentication_url("/redirect"))
+    actual = URLObject(api_client.authentication_url("/redirect", scopes="email"))
     assert urls_equal(expected, actual)
 
     actual2 = URLObject(
-        api_client.authentication_url("/redirect", login_hint="hint")
+        api_client.authentication_url("/redirect", scopes=["calendar", "contacts"])
     )
-    expected2 = expected.set_query_param("login_hint", "hint")
+    expected2 = expected.set_query_param("scopes", "calendar,contacts")
     assert urls_equal(expected2, actual2)
-
-    actual3 = URLObject(
-        api_client.authentication_url("/redirect", state="confusion")
-    )
-    expected3 = expected.set_query_param("state", "confusion")
-    assert urls_equal(expected3, actual3)
 
 
 def test_client_token_for_code(mocked_responses, api_client, api_url):